Make color_set use 65k pairs instead of max(short)
Initially, this code uses the `pair` argument in `color_set(pair, opts)`
to set color_pairs. Since `pair` is of type (_signed_) short, it can only
support 2^15 color_pairs at a time. The terminal supports 2^16 color
pairs, as can be seen by `COLORS` when using `TERM=xterm-direct`. We can
use the entirety of this by passing the address of an `int` pair to
opts, which is no longer limited to the limits of `short`.