Use half block elements as pixels for +resolution
Before this commit, the code used one character per pixel (just with the
same fore/background colors). As characters are taller than they are
wide, the resulting image is stretched vertically, and wastes vertical
resolution. This limitation was solved by using the upper half block element
[0], so that each character corresponds to two pixels (top and bottom),
where the foreground and background corresponds to the color of the top
and bottom pixels respectively. This doubles the vertical resolution,
and removes the vertical stretch.

[0]: https://en.wikipedia.org/wiki/Block_Elements